Enzyme installed. SearchBar moved to its own folder and tests added
