+++ package.json
@@ -12,6 +12,8 @@
     "@testing-library/react": "^10.4.9",
     "@testing-library/user-event": "^12.1.3",
     "clsx": "^1.1.1",
+    "enzyme": "^3.11.0",
+    "enzyme-adapter-react-16": "^1.15.6",
     "moment": "^2.29.1",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
